requirements.txt
streamlit
langchain-community
langchain_groq
faiss-cpu
python-dotenv
PyMuPDF
sentence-transformers
